moved all fundamental array ops to their own file, internal
