+++ Engine/entry.c
@@ -272,6 +272,312 @@ OENTRY opcodlst_1[] = {
   { "##subin.k", S(ASSIGN),0,        "k",    "k",    NULL,   subin   },
   { "##subin.K", S(ASSIGN),0,        "a",    "k",    NULL,   subinak },
   { "##subin.a", S(ASSIGN),0,        "a",    "a",    NULL,   subina  },
+ { "init.i", sizeof(ARRAYINIT), 0, "i[]", "m", (SUBR)array_init },
+    { "init.k", sizeof(ARRAYINIT), 0, "k[]", "m", (SUBR)array_init },
+    { "init.a", sizeof(ARRAYINIT), 0, "a[]", "m", (SUBR)array_init },
+    { "init.S", sizeof(ARRAYINIT), 0, "S[]", "m", (SUBR)array_init },
+    { "init.0", sizeof(ARRAYINIT), 0, ".[]", "m", (SUBR)array_init },
+    { "fillarray.k", sizeof(TABFILL), 0, "k[]", "m", (SUBR)tabfill },
+    { "fillarray.i", sizeof(TABFILL), 0, "i[]", "m", (SUBR)tabfill },
+    { "fillarray.s", sizeof(TABFILL), 0, "S[]", "W", (SUBR)tabfill },
+    { "fillarray.K", sizeof(TABFILL), 0, "k[]", "z", NULL, (SUBR)tabfill },
+    { "fillarray.f", sizeof(TABFILLF), 0, "k[]", "S", (SUBR)tabfillf },
+    { "fillarray.F", sizeof(TABFILLF), 0, "i[]", "S", (SUBR)tabfillf },
+    { "string2array.S", sizeof(TABFILLF), 0, "i[]", "S", (SUBR)tabsfill },
+    { "string2array.s", sizeof(TABFILLF), 0, "k[]", "S", (SUBR)tabsfill },
+    { "array.k", sizeof(TABFILL), _QQ, "k[]", "m", (SUBR)tabfill     },
+    { "array.i", sizeof(TABFILL), _QQ, "i[]", "m", (SUBR)tabfill     },
+    { "array.a", sizeof(A2ARR), 0, "k[]", "a", (SUBR)asig2array_init,
+      (SUBR)asig2array_perf},
+    { "a.A", sizeof(ARR2A), 0, "a", "k[]", NULL, (SUBR)array2asig_perf},
+    { "##array_init", sizeof(ARRAY_SET), 0, "", ".[].m", (SUBR)array_set },
+    { "##array_set.k", sizeof(ARRAY_SET), 0, "", "k[]km", NULL,(SUBR)array_set},
+    { "##array_set.k", sizeof(ARRAY_SET), 0, "", "k[][]km", NULL,(SUBR)array_set},
+    { "##array_set.a", sizeof(ARRAY_SET), 0, "", "a[]am", NULL, (SUBR)array_set},
+    { "##array_set.S", sizeof(ARRAY_SET), 0, "", "S[].m", (SUBR)array_set , (SUBR)array_set },
+    { "##array_set.i", sizeof(ARRAY_SET), 0, "", ".[].m", (SUBR)array_set },
+    { "##array_set.C", sizeof(ARRAY_SET), 0, "", ":Complex;[]Complex;m", NULL, (SUBR)array_set},
+    { "##array_set.e", sizeof(ARRAY_SET), 0, "", "i[].z", (SUBR)array_err },
+    { "##array_set.x", sizeof(ARRAY_SET), 0, "", ".[].z", NULL, (SUBR)array_set},
+    { "##array_get.k", sizeof(ARRAY_GET), 0, "k", "k[]m", NULL,(SUBR)array_get },
+    { "##array_get.a", sizeof(ARRAY_GET), 0, "a", "a[]m",NULL, (SUBR)array_get },
+    { "##array_get.C", sizeof(ARRAY_GET), 0, ":Complex;", ":Complex;[]m", NULL, (SUBR)array_get},
+    { "##array_get.x", sizeof(ARRAY_GET), 0, ".", ".[]m",(SUBR)array_get },
+    { "##array_get.K", sizeof(ARRAY_GET), 0, ".", ".[]z",NULL, (SUBR)array_get},
+    { "i.Ai", sizeof(ARRAY_GET),0,      "i",    "k[]m", (SUBR)array_get  },
+    { "i.Ak", sizeof(ARRAY_GET),0,      "i",    "k[]z", (SUBR)array_get  },
+    {"=.X", sizeof(TABCOPY), 0, "k[]", "k", (SUBR) scalarset, (SUBR) scalarset},
+    {"=.Z", sizeof(TABCOPY), 0 , "a[]", "a", NULL, (SUBR) arrayass},
+    /* ******************************************** */
+    {"##add.[s]", sizeof(TABARITH), 0, "a[]", "a[]a[]",
+     (SUBR)tabarithset, (SUBR)tabaadd},
+    {"##add.[]", sizeof(TABARITH), 0, "k[]", "k[]k[]",
+     (SUBR)tabarithset, (SUBR)tabadd},
+    {"##add.[i]", sizeof(TABARITH), 0, "i[]", "i[]i[]",
+     (SUBR)tabaddi},   
+    /* ******************************************** */
+    {"##sub.[a]", sizeof(TABARITH), 0, "a[]", "a[]a[]",
+     (SUBR)tabarithset, (SUBR)tabasub},
+    {"##sub.[]", sizeof(TABARITH), 0, "k[]", "k[]k[]",
+     (SUBR)tabarithset, (SUBR)tabsub},
+    {"##sub.[i]", sizeof(TABARITH), 0, "i[]", "i[]i[]",
+     (SUBR)tabsubi},
+    {"##mul.[a]", sizeof(TABARITH), 0, "a[]", "a[]a[]",
+     (SUBR)tabarithset,(SUBR)tabamul},
+    {"##mul.[]", sizeof(TABARITH), 0, "k[]", "k[]k[]",
+     (SUBR)tabarithset,(SUBR)tabmult},
+    {"##mul.[i]", sizeof(TABARITH), 0, "i[]", "i[]i[]",
+     (SUBR)tabmulti},
+    {"##div.[a]", sizeof(TABARITH), 0, "a[]", "a[]a[]",
+     (SUBR)tabarithset, (SUBR)tabadiv},
+    {"##div.[]",  sizeof(TABARITH), 0, "k[]", "k[]k[]",
+     (SUBR)tabarithset,(SUBR)tabdiv },
+    {"##div.[i]",  sizeof(TABARITH), 0, "i[]", "i[]i[]",
+     (SUBR)tabdivi },
+    {"##rem.[]",  sizeof(TABARITH), 0, "k[]", "k[]k[]",
+     (SUBR)tabarithset, (SUBR)tabrem},
+    {"##rem.[i]",  sizeof(TABARITH), 0, "i[]", "i[]i[]", (SUBR)tabremi},
+    {"##add.[i", sizeof(TABARITH1), 0, "k[]", "k[]i",
+     (SUBR)tabarithset1, (SUBR)tabaiadd },
+    {"##add.i[", sizeof(TABARITH2), 0, "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiaadd },
+    {"##add.[p", sizeof(TABARITH1), 0, "i[]", "i[]i", (SUBR)tabaiaddi },
+    {"##add.p[", sizeof(TABARITH2), 0, "i[]", "ii[]", (SUBR)tabiaaddi },
+    {"##add.k[a[", sizeof(TABARITH), 0, "a[]", "k[]a[]",
+     (SUBR)tabarithset, (SUBR)tabkrardd },
+    {"##add.a[k[", sizeof(TABARITH), 0, "a[]", "a[]k[]",
+     (SUBR)tabarithset, (SUBR)tabarkrdd },
+    {"##addin.[i", sizeof(TABARITHIN1), 0, "i[]", "i", (SUBR)addinAA, NULL },
+    {"##addin.[k", sizeof(TABARITHIN1), 0, "k[]", "i", NULL, (SUBR)addinAA },
+    {"##addin.[", sizeof(TABARITHIN), 0, "i[]", "i[]",  (SUBR)tabaddinkk, NULL },
+    {"##addin.[K", sizeof(TABARITHIN), 0, "k[]", "k[]", NULL, (SUBR)tabaddinkk },
+    {"##addin.[a", sizeof(TABARITHIN), 0, "a[]", "a[]", NULL, (SUBR)tabaaddin },
+    {"##addin.[a", sizeof(TABARITHIN), 0, "a[]", "a[]", NULL, (SUBR)tabaaddin },
+    {"##addin.[ak", sizeof(TABARITHIN), 0, "a[]", "k[]", NULL, (SUBR)tabarkrddin },
+    {"##addin.[aks", sizeof(TABARITHIN), 0, "a[]", "k", NULL, (SUBR)tabakaddin },
+    {"##sub.[i", sizeof(TABARITH1), 0, "k[]", "k[]i",
+     (SUBR)tabarithset1, (SUBR)tabaisub },
+    {"##sub.i[", sizeof(TABARITH2), 0, "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiasub },
+    {"##subin.[i", sizeof(TABARITHIN1), 0, "i[]", "i", (SUBR)subinAA, NULL },
+    {"##subin.[", sizeof(TABARITHIN), 0, "i[]", "i[]",  (SUBR)tabsubinkk, NULL },
+    {"##subin.[K", sizeof(TABARITHIN), 0, "k[]", "k[]", NULL, (SUBR)tabsubinkk },
+    {"##subin.[k", sizeof(TABARITHIN1), 0, "k[]", "k", NULL, (SUBR)subinAA },
+    {"##subin.[a", sizeof(TABARITHIN), 0, "a[]", "a[]", NULL, (SUBR)tabaasubin },
+    {"##subin.[ak", sizeof(TABARITHIN), 0, "a[]", "k[]", NULL, (SUBR)tabarkrsbin },
+    {"##subdin.[aks", sizeof(TABARITHIN), 0, "a[]", "k", NULL, (SUBR)tabaksubin },
+    {"##sub.[p", sizeof(TABARITH1), 0, "i[]", "i[]i", (SUBR)tabaisubi },
+    {"##sub.p[", sizeof(TABARITH2), 0, "i[]", "ii[]", (SUBR)tabiasubi },
+    {"##sub.k[a[", sizeof(TABARITH), 0, "a[]", "k[]a[]",
+     (SUBR)tabarithset, (SUBR)tabkrarsb },
+    {"##sub.a[k[", sizeof(TABARITH), 0, "a[]", "a[]k[]",
+     (SUBR)tabarithset, (SUBR)tabarkrsb },
+    {"##mul.[i", sizeof(TABARITH1), 0, "k[]", "k[]i",
+     (SUBR)tabarithset1, (SUBR)tabaimult },
+    {"##mul.i[", sizeof(TABARITH2), 0, "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiamult },
+    {"##mul.[p", sizeof(TABARITH1), 0, "i[]", "i[]i", (SUBR)tabaimulti },
+    {"##mul.p[", sizeof(TABARITH2), 0, "i[]", "ii[]",  (SUBR)tabiamulti },
+    {"##mul.k[a[", sizeof(TABARITH), 0, "a[]", "k[]a[]",
+     (SUBR)tabarithset, (SUBR)tabkrarml },
+    {"##mul.a[k[", sizeof(TABARITH), 0, "a[]", "a[]k[]",
+     (SUBR)tabarithset, (SUBR)tabarkrml },
+    {"##div.[i",  sizeof(TABARITH1), 0, "k[]", "k[]i",
+     (SUBR)tabarithset1, (SUBR)tabaidiv },
+    {"##div.i[",  sizeof(TABARITH2), 0, "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiadiv },
+    {"##div.[p",  sizeof(TABARITH1), 0, "i[]", "i[]i", (SUBR)tabaidivi },
+    {"##div.p[",  sizeof(TABARITH2), 0, "i[]", "ii[]", (SUBR)tabiadivi },
+    {"##div.k[a[", sizeof(TABARITH), 0, "a[]", "k[]a[]",
+     (SUBR)tabarithset, (SUBR)tabkrardv },
+    {"##div.a[k[", sizeof(TABARITH), 0, "a[]", "a[]k[]",
+     (SUBR)tabarithset, (SUBR)tabarkrdv },
+    {"##rem.[i",  sizeof(TABARITH1),0,  "k[]", "k[]i",
+     (SUBR)tabarithset1, (SUBR)tabairem },
+    {"##rem.i[",  sizeof(TABARITH2),0,  "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiarem },
+    {"##rem.[p",  sizeof(TABARITH1),0,  "i[]", "i[]i", (SUBR)tabairemi },
+    {"##rem.p[",  sizeof(TABARITH2),0,  "i[]", "ii[]", (SUBR)tabiaremi },
+    {"##rem.a[k[", sizeof(TABARITH), 0, "a[]", "a[]k[]",
+     (SUBR)tabarithset, (SUBR)tabarkrmd },
+    {"##re.k[a[", sizeof(TABARITH), 0, "a[]", "k[]a[]",
+     (SUBR)tabarithset, (SUBR)tabkrarmd },
+    {"##add.[k", sizeof(TABARITH1), 0, "k[]", "i[]k",
+     (SUBR)tabarithset1, (SUBR)tabaiadd },
+    {"##add.[k", sizeof(TABARITH1), 0, "k[]", "k[]k",
+     (SUBR)tabarithset1, (SUBR)tabaiadd },
+    {"##add.k[", sizeof(TABARITH2), 0, "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiaadd },
+    {"##add.k[", sizeof(TABARITH2), 0, "k[]", "kk[]",
+     (SUBR)tabarithset2, (SUBR)tabiaadd },
+    {"##add.[ak", sizeof(TABARITH1), 0, "a[]", "a[]k",
+     (SUBR)tabarithset1, (SUBR)tabakadd },
+    {"##add.k[a", sizeof(TABARITH2), 0, "a[]", "ka[]",
+     (SUBR)tabarithset2, (SUBR)tabkaadd },
+    {"##add.aa[", sizeof(TABARITH1), 0, "a[]", "aa[]",
+     (SUBR)tabarithset1, (SUBR)tabaardd },
+    {"##add.a[a", sizeof(TABARITH2), 0, "a[]", "a[]a",
+     (SUBR)tabarithset2, (SUBR)tabaradd },
+    {"##sub.[k", sizeof(TABARITH1), 0, "k[]", "i[]k",
+     (SUBR)tabarithset1, (SUBR)tabaisub },
+    {"##sub.[k", sizeof(TABARITH1), 0, "k[]", "k[]k",
+     (SUBR)tabarithset1, (SUBR)tabaisub },
+    {"##sub.k[", sizeof(TABARITH2), 0, "k[]", "kk[]",
+     (SUBR)tabarithset2, (SUBR)tabiasub },
+    {"##sub.k[", sizeof(TABARITH2), 0, "k[]", "ki[]",
+     (SUBR)tabarithset2, (SUBR)tabiasub },
+    {"##sub.[ak", sizeof(TABARITH1), 0, "a[]", "a[]k",
+     (SUBR)tabarithset1, (SUBR)tabaksub },
+    {"##sub.k[a", sizeof(TABARITH2), 0, "a[]", "ka[]",
+     (SUBR)tabarithset2, (SUBR)tabkasub },
+    {"##sub.aa[", sizeof(TABARITH2), 0, "a[]", "a[]a",
+     (SUBR)tabarithset2, (SUBR)tabaarsb },
+    {"##sub.a[a", sizeof(TABARITH1), 0, "a[]", "aa[]",
+     (SUBR)tabarithset1, (SUBR)tabarasb },
+    {"##mul.[k", sizeof(TABARITH1), 0, "k[]", "k[]k",
+     (SUBR)tabarithset1, (SUBR)tabaimult },
+    {"##mul.[k", sizeof(TABARITH1), 0, "k[]", "i[]k",
+     (SUBR)tabarithset1, (SUBR)tabaimult },
+    {"##mul.k[", sizeof(TABARITH2), 0, "k[]", "kk[]",
+     (SUBR)tabarithset2, (SUBR)tabiamult },
+    {"##mul.k[", sizeof(TABARITH2), 0, "k[]", "ki[]",
+     (SUBR)tabarithset2, (SUBR)tabiamult },
+    {"##mul.[ak", sizeof(TABARITH1), 0, "a[]", "a[]k",
+     (SUBR)tabarithset1, (SUBR)tabakmult },
+    {"##mul.k[a", sizeof(TABARITH2), 0, "a[]", "ka[]",
+     (SUBR)tabarithset2, (SUBR)tabkamult },
+    {"##mul.a[a", sizeof(TABARITH1), 0, "a[]", "a[]a",
+     (SUBR)tabarithset1, (SUBR)tabaraml },
+    {"##mul.aa[", sizeof(TABARITH2), 0, "a[]", "aa[]",
+     (SUBR)tabarithset2, (SUBR)tabaarml },
+    {"##div.[k",  sizeof(TABARITH1), 0, "k[]", "k[]k",
+     (SUBR)tabarithset1, (SUBR)tabaidiv },
+    {"##div.k[",  sizeof(TABARITH2), 0, "k[]", "ki[]",
+     (SUBR)tabarithset2, (SUBR)tabiadiv },
+    {"##div.[k",  sizeof(TABARITH1), 0, "k[]", "i[]k",
+     (SUBR)tabarithset1, (SUBR)tabaidiv },
+    {"##div.k[",  sizeof(TABARITH2), 0, "k[]", "kk[]",
+     (SUBR)tabarithset2, (SUBR)tabiadiv },
+    {"##div.[ak", sizeof(TABARITH1), 0, "a[]", "a[]k",
+     (SUBR)tabarithset1, (SUBR)tabakdiv },
+    {"##div.k[a", sizeof(TABARITH2), 0, "a[]", "ka[]",
+     (SUBR)tabarithset2, (SUBR)tabkadiv },
+    {"##div.a[a", sizeof(TABARITH1), 0, "a[]", "a[]a",
+     (SUBR)tabarithset1, (SUBR)tabaardv },
+    {"##div.a[a", sizeof(TABARITH1), 0, "a[]", "a[]a",
+     (SUBR)tabarithset1, (SUBR)tabaradv },
+    {"##rem.[k",  sizeof(TABARITH1),0,  "k[]", "k[]k",
+     (SUBR)tabarithset1, (SUBR)tabairem },
+    {"##rem.[ak",  sizeof(TABARITH1),0,  "a[]", "a[]k",
+     (SUBR)tabarithset1, (SUBR)tabarkrem },
+    {"##rem.k[",  sizeof(TABARITH2),0,  "k[]", "kk[]",
+     (SUBR)tabarithset2, (SUBR)tabiarem },
+    {"##pow.[]",  sizeof(TABARITH), 0, "k[]", "k[]k[]",
+     (SUBR)tabarithset,(SUBR)tabpow },
+    {"##pow.[i]",  sizeof(TABARITH), 0, "i[]", "i[]i[]",
+     (SUBR)tabpowi },
+    {"##pow.[i",  sizeof(TABARITH1), 0, "k[]", "k[]i",
+     (SUBR)tabarithset1, (SUBR)tabaipow },
+    {"##pow.i[",  sizeof(TABARITH2), 0, "k[]", "ik[]",
+     (SUBR)tabarithset2, (SUBR)tabiapow },
+    {"##pow.[p",  sizeof(TABARITH1), 0, "i[]", "i[]i", (SUBR)tabaipowi },
+    {"##pow.p[",  sizeof(TABARITH2), 0, "i[]", "ii[]", (SUBR)tabiapowi },
+    {"##pow.[k",  sizeof(TABARITH1), 0, "k[]", "k[]k",
+     (SUBR)tabarithset1, (SUBR)tabaipow },
+    {"##pow.k[",  sizeof(TABARITH2), 0, "k[]", "kk[]",
+     (SUBR)tabarithset2, (SUBR)tabiapow },
+    {"##pow.ki[",  sizeof(TABARITH2), 0, "k[]", "ki[]",
+     (SUBR)tabarithset2, (SUBR)tabiapow },
+    {"##pow.[ak",  sizeof(TABARITH1), 0, "a[]", "a[]k",
+     (SUBR)tabarithset1, (SUBR)tabarkpow },
+    {"##pow.a[k[", sizeof(TABARITH), 0, "a[]", "a[]k[]",
+     (SUBR)tabarithset, (SUBR)tabarkrpw },
+    { "taninvx2.Ai", sizeof(TABARITH), 0, "i[]", "i[]i[]", (SUBR)taninv2_Ai  },
+    { "taninv2.Ak", sizeof(TABARITH), 0, "k[]", "k[]k[]", (SUBR)tabarithset, (SUBR)taninv2_A  },
+    { "taninv2.Aa", sizeof(TABARITH), 0, "a[]", "a[]a[]", (SUBR)tabarithset, (SUBR)taninv2_Aa  },
+    { "clear", sizeof(TABCLEAR), 0, "", "a[]", (SUBR)tabclearset, (SUBR)tabclear  },
+  { "clear", sizeof(TABCLEAR), 0, "", "k[]", (SUBR)tabclearset, (SUBR)tabcleark  },
+    { "maxtab.k",sizeof(TABQUERY),_QQ, "kz", "k[]",
+      (SUBR) tabqset, (SUBR) tabmax },
+    { "maxarray.k", sizeof(TABQUERY), 0, "kz", "k[]",
+      (SUBR) tabqset,(SUBR) tabmax },
+    { "maxarray.i", sizeof(TABQUERY), 0, "iI", "i[]",(SUBR) tabmax1, NULL  },
+    { "mintab.k", sizeof(TABQUERY),_QQ, "kz", "k[]",
+      (SUBR) tabqset, (SUBR) tabmin },
+    { "minarray.k", sizeof(TABQUERY),0, "kz", "k[]",(SUBR) tabqset,
+      (SUBR) tabmin },
+    { "minarray.i", sizeof(TABQUERY),0, "iI", "i[]",(SUBR) tabmin1 },
+    { "sumtab", sizeof(TABQUERY1),_QQ, "k", "k[]",
+      (SUBR) tabqset1, (SUBR) tabsum },
+    { "sumarray.k", sizeof(TABQUERY1),0, "k", "k[]",
+      (SUBR) tabqset1, (SUBR) tabsum },
+    { "sumarray.a", sizeof(TABQUERY1),0, "a", "a[]",
+      (SUBR) tabqset1, (SUBR) tabsuma1 },
+    { "sumarray.i", sizeof(TABQUERY1),0, "i", "i[]", (SUBR) tabsum1   },
+    { "scalet", sizeof(TABSCALE), _QQ|WI, "",  "k[]kkOJ",
+      (SUBR) tabscaleset,(SUBR) tabscale },
+    { "scalearray.k", sizeof(TABSCALE), WI, "",  "k[]kkOJ",
+      (SUBR) tabscaleset,(SUBR) tabscale },
+    { "scalearray.1", sizeof(TABSCALE), WI, "",  "i[]iiOJ",   (SUBR) tabscale1 },
+    { "=.I", sizeof(TABCPY), 0, "i[]", "i[]", (SUBR)tabcopy1, NULL },
+    { "fillarray", sizeof(TABCPY), 0, "k[]", "i[]", (SUBR)tabcopy1, NULL },
+    { "=.J", sizeof(TABCPY), 0, "i[]", "k[]", (SUBR)tabcopy1, NULL },
+    { "=.IK", sizeof(TABCPY), 0, "k[]", "i[]", (SUBR)tabcopy1, (SUBR)tabcopy1 },
+    { "=.K", sizeof(TABCPY), 0, "k[]", "k[]", (SUBR)tabcopyk_init, (SUBR)tabcopyk },
+    { "=._", sizeof(TABCPY), 0, ".[]", ".[]", (SUBR)tabcopy, (SUBR)tabcopy },
+    { "=.L", sizeof(TABCPY), 0, ".[]", ".[]", (SUBR)tabcopy2, (SUBR)tabcopy2 },
+    { "tabgen", sizeof(TABGEN), _QQ, "k[]", "iip", (SUBR) tabgen, NULL    },
+    { "tabmap_i", sizeof(TABMAP), _QQ, "k[]", "k[]S", (SUBR) tabmap_set   },
+    { "tabmap", sizeof(TABMAP), _QQ, "k[]", "k[]S", (SUBR) tabmap_set,
+      (SUBR) tabmap_perf},
+    { "tabmap", sizeof(TABMAP), _QQ, "k[]", "k[]S", (SUBR) tabmap_set,
+      (SUBR) tabmap_perf},
+    { "genarray.i", sizeof(TABGEN),0, "i[]", "iip", (SUBR) tabgen, NULL   },
+    { "genarray_i", sizeof(TABGEN),0, "k[]", "iip", (SUBR) tabgen, NULL, NULL},
+    { "genarray.k", sizeof(TABGEN),0, "k[]", "kkp", NULL, (SUBR)tabgen    },
+    { "maparray_i", sizeof(TABMAP),0, "k[]", "k[]S", (SUBR) tabmap_set    },
+    { "maparray.k", sizeof(TABMAP), 0, "k[]", "k[]S", (SUBR) tabmap_set,
+      (SUBR) tabmap_perf          },
+    { "maparray.i", sizeof(TABMAP), 0, "i[]", "i[]S", (SUBR) tabmap_set },
+    /*  { "maparray.s", sizeof(TABMAP), 0, "S[]", "S[]S", (SUBR) tabmap_set, */
+    /*                                               (SUBR) tabmap_perf     }, */
+    { "tabslice", sizeof(TABSLICE), _QQ, "k[]", "k[]iip",
+      NULL, (SUBR) tabslice, NULL },
+    { "slicearray.i", sizeof(TABSLICE), 0, "i[]", "i[]iip",
+      (SUBR) tabslice, NULL, NULL },
+    { "slicearray.k", sizeof(TABSLICE), 0, "k[]", "k[]iip",
+      (SUBR) tabslice, (SUBR) tabslice, NULL },
+    { "slicearray.a", sizeof(TABSLICE), 0, "a[]", "a[]iip",
+      (SUBR) tabslice, (SUBR) tabslice, NULL },
+    { "slicearray.S", sizeof(TABSLICE), 0, "S[]", "S[]iip",
+      (SUBR) tabslice, (SUBR) tabslice, NULL },
+    { "slicearray_i.i", sizeof(TABSLICE), 0, "i[]", "i[]iip",
+      (SUBR) tabslice, NULL },
+    { "slicearray_i.k", sizeof(TABSLICE), 0, "k[]", "k[]iip",
+      (SUBR) tabslice, NULL },
+    { "slicearray_i.S", sizeof(TABSLICE), 0, "S[]", "S[]iip",
+      (SUBR) tabslice, NULL },
+    { "trim.i", sizeof(TRIM), WI, "", "i[]i", (SUBR)trim_i, NULL },
+    { "trim.k", sizeof(TRIM), WI, "", ".[]k", NULL, (SUBR)trim },
+    { "trim_i", sizeof(TRIM), WI, "", ".[]i", (SUBR)trim_i, NULL },
+    { "copy2ftab", sizeof(TABCOPY), TW|_QQ, "", "k[]k", NULL, (SUBR) tab2ftab },
+    { "copy2ttab", sizeof(TABCOPY), TR|_QQ, "", "k[]k", NULL, (SUBR) ftab2tab },
+    { "copya2ftab.k", sizeof(TABCOPY), TW, "", "k[]k",
+      (SUBR) tab2ftabi, (SUBR) tab2ftab },
+
+    { "copyf2array.k", sizeof(TABCOPY), TR, "", "k[]k",
+      (SUBR) ftab2tabi, (SUBR) ftab2tab },
+    { "copyf2array.kk", sizeof(TABCOPY), TR, "", "k[]kk",
+      (SUBR) ftab2tabi, (SUBR) ftab2tab },
+    // { "copya2ftab.i", sizeof(TABCOPY), TW, "", "i[]i", (SUBR) tab2ftabi },
+    { "copya2ftab.ii", sizeof(TABCOPY2), TW, "", "i[]io", (SUBR) tab2ftab_offset_i },
+    { "copya2ftab.kk", sizeof(TABCOPY2), TW, "", "k[]kO", NULL , (SUBR) tab2ftab_offset },
+    { "copyf2array.i", sizeof(TABCOPY), TR, "", "i[]i", (SUBR) ftab2tabi },
+    /* { "lentab", 0xffff}, */
+    { "lentab.i", sizeof(TABQUERY1), _QQ, "i", "k[]p", (SUBR) tablength },
+    { "lentab.k", sizeof(TABQUERY1), _QQ, "k", "k[]p", NULL, (SUBR) tablength },
+    { "lenarray.ix", sizeof(TABQUERY1), 0, "i", ".[]p", (SUBR) tablength },
+    { "lenarray.kx", sizeof(TABQUERY1), 0, "k", ".[]p", NULL, (SUBR)tablength },
   /* complex arrays */
   {"##add.Complex[]Complex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;[]",
    (SUBR) cops_init,(SUBR)complexa_plus_complexa},
@@ -1433,6 +1739,8 @@ OENTRY opcodlst_1[] = {
     (SUBR)copy_var_generic, NULL},
   { "monitor",  sizeof(MONITOR_OPCODE), IB,   "mmmmmmmmmmmmmmmmmmmmmmmm", "",
     (SUBR) monitor_opcode_init, (SUBR) notinit_opcode_stub,  NULL },
+  { "monitor.A", sizeof(MONITOR_A), IB, "a[]", "",
+      (SUBR)monitora_init, (SUBR)monitora_perf},
   { "outrg", S(OUTRANGE), IR, "", "ky",
     (SUBR)outRange_i, (SUBR)outRange},
   { "nchnls_hw", S(ASSIGN), 0, "ii", "",
