+++ H/entry.h
@@ -67,6 +67,7 @@
 #include "midiops2.h"
 #include "pitch.h"
 #include "opcode.h"
+#include "array_ops.h"
 
 #define S(x)    sizeof(x)
